Link directly to images in the Image Viewer within the category
I find it endlessly annoying when I click on a thumbnail to get a close
up view of an image, and I have to click through several intermediate
pages before I actually get a larger image.

The flow is now perfect (in my humble opinion): after one click, the
image fills the entire page, you can flick to the next image in the
category and the image details are still linked.